v0.1.5 - AI Detection
Now uses a trained AI model to detect keys on MTG cards instead of hard-coded ROIs for OCR

AI was trained on a Raspberry Pi 5 with the only non-stock addition being an NVME HAT

App has fallback logic for alternator layouts. If there is no set code or collector number detected, it will use the set icon if it detects it then compares that icon to a known list of icons. This is still wonky and I'm working on making it better.

App can run on its own with little to no intervention. The odd card gets dropped at an odd angle and needs to be adjusted